refactor: rename searcher.ts to regexps.ts
